:sparkles: Adds Home Tech Hacker blog (#126)
Describe the proposed change
Hi. I know this is self-promotion, but I like to think the information on my blog is useful in helping Home Assistant users and promoting Home Assistant. It has articles on building and improving smarthomes, using and configuring Home Assistant, and reviews of devices that integrate with HA. I will continue to add more of those types of articles.

If it is more appropriate to link directly to my smarthome or Home Assistant category I'm happy to do that. Thank you.

BTW, I've never used Github for anything other viewing other people's projects, so I hope I'm doing all this right!

The link
https://HomeTechHacker.com